Appending empty struct to message_values, fixes segfault in JSCLassCreate

+++ src/util.c
@@ -334,7 +334,7 @@ util_expand_home(char *buffer, const char *filename, size_t length) {
 char * 
 util_normalize_filename(char *buffer, const char *filename, size_t length) {
   char *tmp = buffer;
-  for (int i=0; *filename != 0 && i<length-1; i++, filename++, buffer++) {
+  for (size_t i=0; *filename != 0 && i<length-1; i++, filename++, buffer++) {
     *buffer = *filename == '/' ? '_' : *filename;
   }
   *buffer = 0;
